Permissao de usuario incluido na geraçao do token, CRUD de usuario, model e seeds

+++ src/database/migrations/20200124234633-create-payments.js
@@ -14,7 +14,7 @@ module.exports = {
           allowNull: false
         },
         expiration: {
-          type: Sequelize.INTEGER(2),
+          type: Sequelize.INTEGER,
           allowNull: false
         },
         createdAt: {
